% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{data_loader}
\alias{data_loader}
\title{Data loader}
\usage{
data_loader(dataset_name)
}
\arguments{
\item{dataset_name}{Character. 'ushio_2022_prsb' or 'ushio_2023_elife'.}
}
\description{
\code{data_loader} Batch loading data set.
}
\examples{
# data_loader()
}
