% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_explorer.R
\name{data_to_phyloseq}
\alias{data_to_phyloseq}
\title{Wrapper function to import data.frame to phyloseq}
\usage{
data_to_phyloseq(dataset_name)
}
\arguments{
\item{dataset_name}{Character. 'ushio_2022_prsb', 'ushio_2023a_elife' or 'ushio_2023b_elife'.}
}
\description{
\code{data_to_phyloseq} Import data.frame to phyloseq
}
\examples{
# data_to_phyloseq()
}
