% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2022_prsb_df_climate}
\alias{ushio_2022_prsb_df_climate}
\title{Data from Ushio (2022) Proc. Royal Soc. B: Climate data}
\format{
A data frame with 765 rows and 13 variables:
}
\source{
\url{https://doi.org/10.1098/rspb.2021.2690}
}
\usage{
ushio_2022_prsb_df_climate
}
\description{
Data from Ushio (2022) Proc. Royal Soc. B: Climate data
}
\keyword{datasets}
