% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2022_prsb_df_sample}
\alias{ushio_2022_prsb_df_sample}
\title{Data from Ushio (2022) Proc. Royal Soc. B: Sample data}
\format{
A data frame with 610 rows and 39 variables:
}
\source{
\url{https://doi.org/10.1098/rspb.2021.2690}
}
\usage{
ushio_2022_prsb_df_sample
}
\description{
Data from Ushio (2022) Proc. Royal Soc. B: Sample data
}
\keyword{datasets}
