% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2022_prsb_psobj}
\alias{ushio_2022_prsb_psobj}
\title{Data from Ushio (2022) Proc. Royal Soc. B: A phylooseq object}
\format{
A phyloseq object that contains ASV table, sample information, and taxa information
}
\source{
\url{https://doi.org/10.1098/rspb.2021.2690}
}
\usage{
ushio_2022_prsb_psobj
}
\description{
Data from Ushio (2022) Proc. Royal Soc. B: A phylooseq object
}
\keyword{datasets}
