% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023_elife_df_asvmat}
\alias{ushio_2023_elife_df_asvmat}
\title{Data from Ushio et al. (2023) eLife: ASV table}
\format{
A data frame with 324 rows and 1414 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.87202.1}
}
\usage{
ushio_2023_elife_df_asvmat
}
\description{
Data from Ushio et al. (2023) eLife: ASV table
}
\keyword{datasets}
