% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023_elife_df_rice2019_gr}
\alias{ushio_2023_elife_df_rice2019_gr}
\title{Data from Ushio (2023) eLife: Rice growth monitoring data in 2019}
\format{
A data frame with 405 rows and 16 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.87202.1}
}
\usage{
ushio_2023_elife_df_rice2019_gr
}
\description{
Data from Ushio (2023) eLife: Rice growth monitoring data in 2019
}
\keyword{datasets}
