% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023_elife_df_rice2019_hd}
\alias{ushio_2023_elife_df_rice2019_hd}
\title{Data from Ushio (2023) eLife: Rice heading monitoring data in 2019}
\format{
A data frame with 171 rows and 10 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.87202.1}
}
\usage{
ushio_2023_elife_df_rice2019_hd
}
\description{
Data from Ushio (2023) eLife: Rice heading monitoring data in 2019
}
\keyword{datasets}
