% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023_elife_psobj}
\alias{ushio_2023_elife_psobj}
\title{Data from Ushio (2023) eLife: A phyloseq object}
\format{
A phyloseq object that contains ASV table, sample information, and taxa information
}
\source{
\url{https://doi.org/10.7554/eLife.87202.1}
}
\usage{
ushio_2023_elife_psobj
}
\description{
Data from Ushio (2023) eLife: A phyloseq object
}
\keyword{datasets}
