% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023a_elife_df_asvtab_copy}
\alias{ushio_2023a_elife_df_asvtab_copy}
\title{Data from Ushio et al. (2023a) eLife: ASV table (copy numbers)}
\format{
A data frame with 550 rows and 856 variables:

A data frame with 550 rows and 857 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.85795}

\url{https://doi.org/10.7554/eLife.85795}
}
\usage{
ushio_2023a_elife_df_asvtab_copy

ushio_2023a_elife_df_asvtab_copy
}
\description{
Data from Ushio et al. (2023a) eLife: ASV table (copy numbers)

Data from Ushio et al. (2023a) eLife: ASV table (eDNA copy numbers)
}
\keyword{datasets}
