% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023a_elife_df_asvtab_read}
\alias{ushio_2023a_elife_df_asvtab_read}
\title{Data from Ushio et al. (2023a) eLife: ASV table (sequence reads)}
\format{
A data frame with 550 rows and 856 variables:

A data frame with 550 rows and 857 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.85795}

\url{https://doi.org/10.7554/eLife.85795}
}
\usage{
ushio_2023a_elife_df_asvtab_read

ushio_2023a_elife_df_asvtab_read
}
\description{
Data from Ushio et al. (2023a) eLife: ASV table (sequence reads)

Data from Ushio et al. (2023a) eLife: ASV table (eDNA reads)
}
\keyword{datasets}
