% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023a_elife_df_sample}
\alias{ushio_2023a_elife_df_sample}
\title{Data from Ushio et al. (2023a) eLife: Sample data}
\format{
A data frame with 550 rows and 28 variables:

A data frame with 550 rows and 857 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.85795}

\url{https://doi.org/10.7554/eLife.85795}
}
\usage{
ushio_2023a_elife_df_sample

ushio_2023a_elife_df_sample
}
\description{
Data from Ushio et al. (2023a) eLife: Sample data

Data from Ushio et al. (2023a) eLife: Sample data
}
\keyword{datasets}
