% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023a_elife_df_taxa}
\alias{ushio_2023a_elife_df_taxa}
\title{Data from Ushio et al. (2023a) eLife: Taxa table}
\format{
A data frame with 856 rows and 14 variables:

A data frame with 856 rows and 15 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.85795}

\url{https://doi.org/10.7554/eLife.85795}
}
\usage{
ushio_2023a_elife_df_taxa

ushio_2023a_elife_df_taxa
}
\description{
Data from Ushio et al. (2023a) eLife: Taxa table

Data from Ushio et al. (2023a) eLife: Taxa table
}
\keyword{datasets}
