% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023b_elife_df_climate2019}
\alias{ushio_2023b_elife_df_climate2019}
\title{Data from Ushio (2023b) eLife: Climate data in 2019}
\format{
A data frame with 184 rows and 14 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.87202.1}
}
\usage{
ushio_2023b_elife_df_climate2019
}
\description{
Data from Ushio (2023b) eLife: Climate data in 2019
}
\keyword{datasets}
