% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023b_elife_df_rice2017}
\alias{ushio_2023b_elife_df_rice2017}
\title{Data from Ushio et al. (2023b) eLife: Rice monitoring data in 2017}
\format{
A data frame with 610 rows and 20 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.87202.1}
}
\usage{
ushio_2023b_elife_df_rice2017
}
\description{
Data from Ushio et al. (2023b) eLife: Rice monitoring data in 2017
}
\keyword{datasets}
