% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushio_2023b_elife_df_taxa}
\alias{ushio_2023b_elife_df_taxa}
\title{Data from Ushio et al. (2023b) eLife: Taxa table for 2019}
\format{
A data frame with 1414 rows and 24 variables:
}
\source{
\url{https://doi.org/10.7554/eLife.87202.1}
}
\usage{
ushio_2023b_elife_df_taxa
}
\description{
Data from Ushio et al. (2023b) eLife: Taxa table for 2019
}
\keyword{datasets}
