% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_explorer.R
\name{view_ts_multitax}
\alias{view_ts_multitax}
\title{Visualize multitaxa time series}
\usage{
view_ts_multitax(ps_obj, phy_rank, top_taxa_n = 11, remove_others = TRUE)
}
\arguments{
\item{ps_obj}{\code{phyloseq} object. Recommend using \code{phyloseq} object generated by \code{reed::data_to_phyloseq}.}

\item{phy_rank}{Character. Name of phylogeny rank.}

\item{top_taxa_n}{Numeric. The number of top taxa to be shown.}

\item{remove_others}{Logical. If TRUE, "Others" and "Undetermined" are not shown.}
}
\description{
\code{view_ts_multitax} Visualize multitaxa time series
}
\examples{
# view_ts_multitax()
}
