% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_explorer.R
\name{view_ts_singletax}
\alias{view_ts_singletax}
\title{Visualize single taxon time series}
\usage{
view_ts_singletax(ps_obj, taxon, phy_rank)
}
\arguments{
\item{ps_obj}{\code{phyloseq} object. Recommend using \code{phyloseq} object generated by \code{reed::data_to_phyloseq}.}

\item{taxon}{Character. Taxon name.}

\item{phy_rank}{Character. Name of phylogeny rank.}
}
\description{
\code{view_ts_singletax} Visualize single taxon time series
}
\examples{
# view_ts_singletax()
}
